% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{annotate_genotypes}
\alias{annotate_genotypes}
\title{Annotate genotypes}
\usage{
annotate_genotypes(loom, barcodes, genotypes)
}
\arguments{
\item{loom}{loom file object}

\item{barcodes}{barcodes data}

\item{genotypes}{genotype data}
}
\value{
annotated genotypes
}
\description{
Annotate genotypes
}
\examples{
\dontrun{
genotypes.annotated <- annotate_genotypes(loom, barcodes, genotypes)
}
}
