% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_barcodes}
\alias{cluster_barcodes}
\title{Identify max clusters}
\usage{
cluster_barcodes(bcfile.norm, method, outfile = NULL)
}
\arguments{
\item{bcfile.norm}{Normalised barcode distribution}

\item{method}{Clustering method}

\item{outfile}{Save clusters plot to outfile}
}
\value{
max_cluster
}
\description{
Identify max clusters
}
\examples{
\dontrun{
max_cluster <- cluster_barcodes(bcfile.norm, "kmeans")
max_cluster <- cluster_barcodes(bcfile.norm, "kmeans", "sample-cluster.pdf")
}
}
