% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{compute_ploidy}
\alias{compute_ploidy}
\title{Identify ploidy for samples}
\usage{
compute_ploidy(bcfile.norm, method, max_cluster, clust = NULL,
  norm.cluster = NULL)
}
\arguments{
\item{bcfile.norm}{Normalised barcode distribution}

\item{method}{Clustering method}

\item{max_cluster}{number of clusters}

\item{clust}{barcode clusters based on genotypes}

\item{norm.cluster}{cluster id for normal cluster}
}
\value{
ploidy
}
\description{
Identify ploidy for samples
}
\examples{
\dontrun{
ploidy <- compute_ploidy(bcfile.norm, "kmeans", max_cluster)
ploidy <- compute_ploidy(bcfile.norm, "kmeans", max_cluster, clust, 1)
}

}
