% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loom.R
\name{connect_to_loom}
\alias{connect_to_loom}
\title{Connect to LOOM file}
\usage{
connect_to_loom(loom_file)
}
\arguments{
\item{loom_file}{LOOM file}
}
\value{
loom connection to input LOOM file
}
\description{
Connect to LOOM file
}
\examples{
\dontrun{
loom <- connect_to_loom(loom_file)
}
}
