% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{extract_genotypes}
\alias{extract_genotypes}
\title{Extract genotypes from loom}
\usage{
extract_genotypes(lfile, barcodes, gt.filter = TRUE, gt.gqc = 30,
  gt.dpc = 10, gt.afc = 20, gt.mv = 50, gt.mc = 50, gt.mm = 1,
  gt.mask = FALSE)
}
\arguments{
\item{lfile}{loom file}

\item{barcodes}{normalized barcode distribution data}

\item{gt.filter}{enable genotype quality filter (Enabled by default)}

\item{gt.gqc}{Genotype quality cutoff (default 30)}

\item{gt.dpc}{Read depth cutoff (default 10)}

\item{gt.afc}{Allele frequency cutoff (default 20)}

\item{gt.mv}{Remove variants genotyped in < X percent of cells (default 50)}

\item{gt.mc}{Remove cells with mutations in < X percent of variants (default 50)}

\item{gt.mm}{Remove variants mutated in < X percent of cells (default 1)}

\item{gt.mask}{mask low quality GT as missing (if GQ/DP/AF lower than cutoff, default FALSE)}
}
\value{
loh data
}
\description{
Extract genotypes from loom
}
\examples{
\dontrun{
genotypes <- extract_genotypes(lfile, barcodes, gt.filter = FALSE)
genotypes <- extract_genotypes(lfile, barcodes,
  gt.filter = TRUE, gt.gqc = 30, gt.dpc = 10, gt.afc = 20,
  gt.mv = 50, gt.mc = 50, gt.mm = 1, gt.mask = FALSE
)
}
}
