% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{filter_by_zygosity}
\alias{filter_by_zygosity}
\title{Filter genotypes based on zygpsity fraction}
\usage{
filter_by_zygosity(genotypes, z.variants = 90, z.missing = 100,
  z.het.min = 5)
}
\arguments{
\item{genotypes}{genotype data}

\item{z.variants}{zygosity percentage cutoff (Default 90)}

\item{z.missing}{missing zygosity cutoff (Default 100)}

\item{z.het.min}{minimum het percentage (Default 5)}
}
\value{
filtered genotypes
}
\description{
Filter genotypes based on zygpsity fraction
}
\examples{
\dontrun{
genotypes.filtered <- filter_by_zygosity(genotypes, 90, 100, 5)
}
}
