% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{filter_common_variants}
\alias{filter_common_variants}
\title{Select Distinctive Mutations}
\usage{
filter_common_variants(genotypes, max_cluster = 2, delta.common = 0.01)
}
\arguments{
\item{genotypes}{filtered data}

\item{max_cluster}{number of cluster to be evaluated}

\item{delta.common}{difference to allow variants}
}
\value{
filtered genotypes
}
\description{
Select Distinctive Mutations
}
\examples{
\dontrun{
genotypes.f <- filter_common_variants(genotypes)
}
}
