% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{genotypes_from_allele_frequency}
\alias{genotypes_from_allele_frequency}
\title{Genotypes from allele frequency}
\usage{
genotypes_from_allele_frequency(loh)
}
\arguments{
\item{loh}{Annotated genotypes}
}
\value{
annotated genotypes
}
\description{
Genotypes from allele frequency
}
\examples{
\dontrun{
genotypes <- genotypes_from_allele_frequency(loh)
}
}
