% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{normalize_barcodes}
\alias{normalize_barcodes}
\title{Normalize barcode data}
\usage{
normalize_barcodes(barcodes, advanced = TRUE)
}
\arguments{
\item{barcodes}{data frame containing merged barcodes from all tubes}

\item{advanced}{advance normalization (dividing read counts by cell median before applying normalization by total sum)}
}
\value{
normalized_barcodes - normalized barcode data
}
\description{
Normalize barcode data
}
\examples{
\dontrun{
normalized_barcodes <- normalize_barcodes(barcodes)
}
}
