% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bed.R
\name{parse_bed}
\alias{parse_bed}
\title{Load bed data}
\usage{
parse_bed(bedfile)
}
\arguments{
\item{bedfile}{bed file}
}
\value{
bed records
}
\description{
Load bed data
}
\examples{
\dontrun{
ampl <- parse_bed("Myeloid_v2_amplicon.bed")
targets <- parse_bed("Myeloid_v2_submitted.bed")
}
}
