% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{ploidy_heatmap}
\alias{ploidy_heatmap}
\title{Save heatmap}
\usage{
ploidy_heatmap(ploidy, amplicons, outfile = NULL,
  groupby.genes = FALSE, min.amplicons = 3)
}
\arguments{
\item{ploidy}{ploidy data}

\item{amplicons}{amplicons bed records}

\item{outfile}{outfile to save plots}

\item{groupby.genes}{group ploidy values by genes}

\item{min.amplicons}{remove genes with less than min.amplicons number of amplicons}
}
\description{
Save heatmap
}
\examples{
\dontrun{
ploidy_heatmap(ploidy, amplicons, "ploidy.pdf")
ploidy_heatmap(ploidy, amplicons, "ploidy.pdf", groupby.genes=TRUE, min.amplicons=3)
}
}
