% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ploidy.R
\name{ploidy_lineplot}
\alias{ploidy_lineplot}
\title{Plot line plot of CNV values for each variant}
\usage{
ploidy_lineplot(ploidy, amplicons, outfile = NULL, clust = NULL,
  cluster_id = NULL, min.amplicons = 3)
}
\arguments{
\item{ploidy}{ploidy}

\item{amplicons}{amplicon bed file}

\item{outfile}{name of output PDF for CNV line plot}

\item{clust}{clustering output from snv heatmap}

\item{cluster_id}{normal or control cluster}

\item{min.amplicons}{filter genes with amplicons < min.amplicons}
}
\description{
Plot line plot of CNV values for each variant
}
\examples{
\dontrun{
ploidy_lineplot(ploidy, amplicons, "lineplot.pdf", clust, 1, 3)
}
}
