% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcodes.R
\name{read_barcodes}
\alias{read_barcodes}
\title{Read barcode distribution data}
\usage{
read_barcodes(files, headerfile)
}
\arguments{
\item{files}{List of tab-delimited barcode distribution files output by Tapestri Pipeline}

\item{headerfile}{vcf header file with barcode mapping}
}
\value{
barcodes - data frame containing merged barcode data from all barcode files
}
\description{
Read barcode distribution data
}
\examples{
\dontrun{
barcodes <- read_barcodes(
  c(
    "/path/to/tube-0.barcode.cell.distribution.tsv",
    "/path/to/tube-1.barcode.cell.distribution.tsv"
  )
)
}
}
