% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{snv_clustering}
\alias{snv_clustering}
\title{Get cell cluster labels}
\usage{
snv_clustering(genotypes, max_cluster)
}
\arguments{
\item{genotypes}{filtered gt matrix}

\item{max_cluster}{maximum number of clusters}
}
\value{
cell cluster labels as dataframe
}
\description{
Get cell cluster labels
}
\examples{
\dontrun{
snv_clust <- snv_clustering(genotypes, max_cluster)
}
}
