% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loh.R
\name{snv_heatmap}
\alias{snv_heatmap}
\title{Heatmap for filtered genotype and cells}
\usage{
snv_heatmap(genotypes, max_cluster = NULL, outfile = NULL,
  median.filter = TRUE, median.bin = 11, annotations = NULL)
}
\arguments{
\item{genotypes}{genotypes}

\item{max_cluster}{max_cluster}

\item{outfile}{output file}

\item{median.filter}{apply median filter while drawing heatmap}

\item{median.bin}{smoothing bin size}

\item{annotations}{annotated genotypes dataframe}
}
\value{
clust
}
\description{
Heatmap for filtered genotype and cells
}
\examples{
\dontrun{
clust <- snv_heatmap(genotypes, max_cluster, outfile, TRUE, FALSE)
}
}
